[CP-181] refactor: RoomException 생성자 직접 호출 방식으로 변경
