+++ src/main/java/com/github/classpick/defaultfile/exception/DefaultFileException.java
@@ -1,6 +1,6 @@
 package com.github.classpick.defaultfile.exception;
 
-import com.github.classpick.global.CustomException;
+import com.github.classpick.global.exception.CustomException;
 
 public class DefaultFileException extends CustomException {
     public DefaultFileException(String message, int status) {
