+++ src/main/java/com/github/classpick/reservation/exception/ReservationException.java
@@ -1,6 +1,6 @@
 package com.github.classpick.reservation.exception;
 
-import com.github.classpick.global.CustomException;
+import com.github.classpick.global.exception.CustomException;
 
 public class ReservationException extends CustomException {
     public ReservationException(String message, int status) {
