+++ src/main/java/com/github/classpick/room/exception/RoomException.java
@@ -1,13 +1,9 @@
 package com.github.classpick.room.exception;
 
-import com.github.classpick.global.CustomException;
+import com.github.classpick.global.exception.CustomException;
 
 public class RoomException extends CustomException {
-    public RoomException(RoomExceptions roomExceptions) {
-        super(roomExceptions.getMessage(), roomExceptions.getCode());
-    }
-
-    public static RoomException of(RoomExceptions roomExceptions) {
-        return new RoomException(roomExceptions);
+    public RoomException(String message, int code) {
+        super(message, code);
     }
 }
