+++ src/main/java/com/github/classpick/user/exception/UserException.java
@@ -1,6 +1,6 @@
 package com.github.classpick.user.exception;
 
-import com.github.classpick.global.CustomException;
+import com.github.classpick.global.exception.CustomException;
 
 public class UserException extends CustomException {
     public UserException(String message, int status) {
